modify the controllers to render the json of exact data
