+++ app/controllers/application_controller.rb
@@ -2,7 +2,6 @@ class ApplicationController < ActionController::Base
   protect_from_forgery with: :exception
   before_action :update_allowed_parameters, if: :devise_controller?
   before_action :authenticate_user!
-
   protect_from_forgery with: :null_session,
                        if: proc { |c| c.request.format =~ %r{application/json} }
   protected
